Make rope work with dynamically loaded modules elsewhere.
E.g., on Fedora-derived Linux distributions `time` (which is not in
`sys.builtin_module_names` and it is dynamically loaded from
`lib-dynload/timemodule.so`) is not recognized as part of the standard
library, which makes ImportUtilsTest.test_sorting_imports_and_standard_modules2 fail (because of superfluous \n included between import of `sys` and `time` modules).

If it doesn’t help on other platforms, we can try some more articulated
magic with __import__.

Created with substantial help from `keypusher` on `#python` on Freenode
IRC channel.

Fixes #38