Don't change kwarg names when moving modules.
This fixes the following error:

```python
rope.base.exceptions.ModuleSyntaxError: Syntax Error in file <...> line <...>: keyword can't be an expression
```

This error comes from having a kwarg in a function with the same name as
the module you're moving. This change fixes this by adding a filter to
ignore keyword parameter names, which is used on the move module
refactor.