Add an option to disable sorting imports entirely
Rope, by default, rewrites import statements by removing all the
top-level import lines from the file, then sorting them, then
reinserting them at the top. This behavior of pulling everything to the
top of the file can be undesirable. As an alternative, we can rewrite
the imports in place, which won't sort them, but also won't impose this
extra constraint that all top-level imports must be at the top.