Perform signature verification on pickled data transferred over sockets (#251)
Perform signature verification on pickled data transferred over sockets

Before unpickling anything, ensure that it has a valid digital
signature using a randomly-generated shared key. In order for an attacker
to send or tamper with data on the same socket, they must know this key
to compute a valid signature. Fixes #105, CVE-2015-3539.