Read source_code from pymodule rather than resource.read()
When the project pref `ignore_syntax_errors=True` is set, the source
code of the pymodule is replaced with `"\n"` [0] to ignore it. But, the
refactoring looks at the actual source code of the module and finds
occurrences which the pymodule doesn't think exist, leading to obscure
IndexErrors. This alternative seems safer and all-around better than the
previous way.