Rename `rope.base.{astutils->nameanalyze}`
`astutils` is a very bad module name for this, many parts of static
analysis can be described as "utilities to work with ast", so it's not
really descriptive enough as to what get_name_levels() and
`_NodeNameCollector` really does. The content of this current module
isn't actually generic utilities either, they do something very
specific, and so calling it `astutils` is also very misleading.