Fix syntax error in test code sample
The sample code in the test was not a valid python code, seems like
they were missing a newline. The previous tokenizer didn't mind these,
but they had always been a syntaxerror when parsed properly. The
tokenizer in Python 3.12 didn't like these forms, so we now have to fix
them.